+++ src/basic/product/product.js
@@ -1,4 +1,4 @@
-import { productList } from "./constants";
+import { state } from "../config/constants";
 
 /**
  * 상품 선택 옵션 업데이트
@@ -6,14 +6,14 @@ import { productList } from "./constants";
  * - 재고가 0인 상품은 선택 불가하도록 설정
  */
 export function updatePickProduct() {
-  selectedProduct.innerHTML = ""; // 기존 옵션 제거
+  state.selectedProduct.innerHTML = ""; // 기존 옵션 제거
 
-  productList.forEach((product) => {
+  state.productList.forEach((product) => {
     const option= document.createElement("option");
     option.value = product.id; // 상품 ID 설정
     option.textContent = `${product.name} - ${product.price}원`; // 상품명 및 가격 표시
     option.disabled = product.quantity === 0; // 재고가 0일경우 비활성화
-    selectedProduct.appendChild(option); // 옵션 추가
+    state.selectedProduct.appendChild(option); // 옵션 추가
   });
 }
 
@@ -34,4 +34,13 @@ export function createCartItem(product) {
     </div>
     `
   return item;
+}
+
+export function updateStockInfo() {
+  const lowStockInfo = state.productList
+    .filter((item) => item.quantity < state.constants.LOW_STOCK_THRESHOLD)
+    .map((item) => `${item.name}: ${item.quantity > 0 ? `재고 부족 (${item.quantity}개 남음)` : "품절"}`)
+    .join("\n"); // 배열을 문자열로 변환할 때 각 요소를 줄바꿈 문자로 연결
+
+  state.stockInfo.textContent = lowStockInfo;
 }
